Accommodate both node and browser builds.
