+++ package.json
@@ -56,6 +56,7 @@
     "depcheck": "^0.6.7",
     "eslint": "^4.6.1",
     "eslint-config-google": "^0.9.1",
+    "event-stream": "^3.3.4",
     "express": "^4.15.4",
     "fs-extra": "^4.0.1",
     "geckodriver": "^1.8.1",
